[AIRFLOW-3458] Move models.Connection into separate file (#4335)
