+++ airflow/utils/db.py
@@ -78,9 +78,8 @@ def provide_session(func):
 
 @provide_session
 def merge_conn(conn, session=None):
-    from airflow import models
-    C = models.Connection
-    if not session.query(C).filter(C.conn_id == conn.conn_id).first():
+    from airflow.models.connection import Connection
+    if not session.query(Connection).filter(Connection.conn_id == conn.conn_id).first():
         session.add(conn)
         session.commit()
 
@@ -89,133 +88,134 @@ def initdb(rbac=False):
     session = settings.Session()
 
     from airflow import models
+    from airflow.models.connection import Connection
     upgradedb()
 
     merge_conn(
-        models.Connection(
+        Connection(
             conn_id='airflow_db', conn_type='mysql',
             host='mysql', login='root', password='',
             schema='airflow'))
     merge_conn(
-        models.Connection(
+        Connection(
             conn_id='beeline_default', conn_type='beeline', port="10000",
             host='localhost', extra="{\"use_beeline\": true, \"auth\": \"\"}",
             schema='default'))
     merge_conn(
-        models.Connection(
+        Connection(
             conn_id='bigquery_default', conn_type='google_cloud_platform',
             schema='default'))
     merge_conn(
-        models.Connection(
+        Connection(
             conn_id='local_mysql', conn_type='mysql',
             host='localhost', login='airflow', password='airflow',
             schema='airflow'))
     merge_conn(
-        models.Connection(
+        Connection(
             conn_id='presto_default', conn_type='presto',
             host='localhost',
             schema='hive', port=3400))
     merge_conn(
-        models.Connection(
+        Connection(
             conn_id='google_cloud_default', conn_type='google_cloud_platform',
             schema='default',))
     merge_conn(
-        models.Connection(
+        Connection(
             conn_id='hive_cli_default', conn_type='hive_cli',
             schema='default',))
     merge_conn(
-        models.Connection(
+        Connection(
             conn_id='hiveserver2_default', conn_type='hiveserver2',
             host='localhost',
             schema='default', port=10000))
     merge_conn(
-        models.Connection(
+        Connection(
             conn_id='metastore_default', conn_type='hive_metastore',
             host='localhost', extra="{\"authMechanism\": \"PLAIN\"}",
             port=9083))
     merge_conn(
-        models.Connection(
+        Connection(
             conn_id='mongo_default', conn_type='mongo',
             host='mongo', port=27017))
     merge_conn(
-        models.Connection(
+        Connection(
             conn_id='mysql_default', conn_type='mysql',
             login='root',
             schema='airflow',
             host='mysql'))
     merge_conn(
-        models.Connection(
+        Connection(
             conn_id='postgres_default', conn_type='postgres',
             login='postgres',
             password='airflow',
             schema='airflow',
             host='postgres'))
     merge_conn(
-        models.Connection(
+        Connection(
             conn_id='sqlite_default', conn_type='sqlite',
             host='/tmp/sqlite_default.db'))
     merge_conn(
-        models.Connection(
+        Connection(
             conn_id='http_default', conn_type='http',
             host='https://www.google.com/'))
     merge_conn(
-        models.Connection(
+        Connection(
             conn_id='mssql_default', conn_type='mssql',
             host='localhost', port=1433))
     merge_conn(
-        models.Connection(
+        Connection(
             conn_id='vertica_default', conn_type='vertica',
             host='localhost', port=5433))
     merge_conn(
-        models.Connection(
+        Connection(
             conn_id='wasb_default', conn_type='wasb',
             extra='{"sas_token": null}'))
     merge_conn(
-        models.Connection(
+        Connection(
             conn_id='webhdfs_default', conn_type='hdfs',
             host='localhost', port=50070))
     merge_conn(
-        models.Connection(
+        Connection(
             conn_id='ssh_default', conn_type='ssh',
             host='localhost'))
     merge_conn(
-        models.Connection(
+        Connection(
             conn_id='sftp_default', conn_type='sftp',
             host='localhost', port=22, login='airflow',
             extra='''
                 {"key_file": "~/.ssh/id_rsa", "no_host_key_check": true}
             '''))
     merge_conn(
-        models.Connection(
+        Connection(
             conn_id='fs_default', conn_type='fs',
             extra='{"path": "/"}'))
     merge_conn(
-        models.Connection(
+        Connection(
             conn_id='aws_default', conn_type='aws',
             extra='{"region_name": "us-east-1"}'))
     merge_conn(
-        models.Connection(
+        Connection(
             conn_id='spark_default', conn_type='spark',
             host='yarn', extra='{"queue": "root.default"}'))
     merge_conn(
-        models.Connection(
+        Connection(
             conn_id='druid_broker_default', conn_type='druid',
             host='druid-broker', port=8082, extra='{"endpoint": "druid/v2/sql"}'))
     merge_conn(
-        models.Connection(
+        Connection(
             conn_id='druid_ingest_default', conn_type='druid',
             host='druid-overlord', port=8081, extra='{"endpoint": "druid/indexer/v1/task"}'))
     merge_conn(
-        models.Connection(
+        Connection(
             conn_id='redis_default', conn_type='redis',
             host='redis', port=6379,
             extra='{"db": 0}'))
     merge_conn(
-        models.Connection(
+        Connection(
             conn_id='sqoop_default', conn_type='sqoop',
             host='rmdbs', extra=''))
     merge_conn(
-        models.Connection(
+        Connection(
             conn_id='emr_default', conn_type='emr',
             extra='''
                 {   "Name": "default_job_flow_name",
@@ -262,27 +262,27 @@ def initdb(rbac=False):
                 }
             '''))
     merge_conn(
-        models.Connection(
+        Connection(
             conn_id='databricks_default', conn_type='databricks',
             host='localhost'))
     merge_conn(
-        models.Connection(
+        Connection(
             conn_id='qubole_default', conn_type='qubole',
             host='localhost'))
     merge_conn(
-        models.Connection(
+        Connection(
             conn_id='segment_default', conn_type='segment',
             extra='{"write_key": "my-segment-write-key"}')),
     merge_conn(
-        models.Connection(
+        Connection(
             conn_id='azure_data_lake_default', conn_type='azure_data_lake',
             extra='{"tenant": "<TENANT>", "account_name": "<ACCOUNTNAME>" }'))
     merge_conn(
-        models.Connection(
+        Connection(
             conn_id='azure_cosmos_default', conn_type='azure_cosmos',
             extra='{"database_name": "<DATABASE_NAME>", "collection_name": "<COLLECTION_NAME>" }'))
     merge_conn(
-        models.Connection(
+        Connection(
             conn_id='cassandra_default', conn_type='cassandra',
             host='cassandra', port=9042))
