+++ airflow/www_rbac/forms.py
@@ -22,7 +22,7 @@ from __future__ import division
 from __future__ import print_function
 from __future__ import unicode_literals
 
-from airflow import models
+from airflow.models.connection import Connection
 from airflow.utils import timezone
 
 from flask_appbuilder.forms import DynamicForm
@@ -93,7 +93,7 @@ class ConnectionForm(DynamicForm):
         widget=BS3TextFieldWidget())
     conn_type = SelectField(
         lazy_gettext('Conn Type'),
-        choices=models.Connection._types,
+        choices=Connection._types,
         widget=Select2Widget())
     host = StringField(
         lazy_gettext('Host'),
