+++ tests/contrib/hooks/test_azure_fileshare_hook.py
@@ -23,8 +23,8 @@ import json
 import unittest
 
 from airflow import configuration
-from airflow import models
 from airflow.contrib.hooks.azure_fileshare_hook import AzureFileShareHook
+from airflow.models.connection import Connection
 from airflow.utils import db
 
 
@@ -42,13 +42,13 @@ class TestAzureFileshareHook(unittest.TestCase):
     def setUp(self):
         configuration.load_test_config()
         db.merge_conn(
-            models.Connection(
+            Connection(
                 conn_id='wasb_test_key', conn_type='wasb',
                 login='login', password='key'
             )
         )
         db.merge_conn(
-            models.Connection(
+            Connection(
                 conn_id='wasb_test_sas_token', conn_type='wasb',
                 login='login', extra=json.dumps({'sas_token': 'token'})
             )
