+++ tests/contrib/hooks/test_databricks_hook.py
@@ -31,7 +31,7 @@ from airflow.contrib.hooks.databricks_hook import (
     SUBMIT_RUN_ENDPOINT
 )
 from airflow.exceptions import AirflowException
-from airflow.models import Connection
+from airflow.models.connection import Connection
 from airflow.utils import db
 
 try:
