+++ tests/contrib/hooks/test_spark_sql_hook.py
@@ -24,7 +24,8 @@ from itertools import dropwhile
 
 from mock import patch, call
 
-from airflow import configuration, models
+from airflow import configuration
+from airflow.models.connection import Connection
 from airflow.utils import db
 from airflow.contrib.hooks.spark_sql_hook import SparkSqlHook
 
@@ -53,7 +54,7 @@ class TestSparkSqlHook(unittest.TestCase):
 
         configuration.load_test_config()
         db.merge_conn(
-            models.Connection(
+            Connection(
                 conn_id='spark_default', conn_type='spark',
                 host='yarn://yarn-master')
         )
