+++ tests/contrib/hooks/test_wasb_hook.py
@@ -24,8 +24,8 @@ import unittest
 from collections import namedtuple
 
 from airflow import configuration, AirflowException
-from airflow import models
 from airflow.contrib.hooks.wasb_hook import WasbHook
+from airflow.models.connection import Connection
 from airflow.utils import db
 
 try:
@@ -42,13 +42,13 @@ class TestWasbHook(unittest.TestCase):
     def setUp(self):
         configuration.load_test_config()
         db.merge_conn(
-            models.Connection(
+            Connection(
                 conn_id='wasb_test_key', conn_type='wasb',
                 login='login', password='key'
             )
         )
         db.merge_conn(
-            models.Connection(
+            Connection(
                 conn_id='wasb_test_sas_token', conn_type='wasb',
                 login='login', extra=json.dumps({'sas_token': 'token'})
             )
