+++ tests/contrib/sensors/test_imap_attachment_sensor.py
@@ -21,8 +21,9 @@ import unittest
 
 from mock import patch, Mock
 
-from airflow import configuration, models
+from airflow import configuration
 from airflow.contrib.sensors.imap_attachment_sensor import ImapAttachmentSensor
+from airflow.models.connection import Connection
 from airflow.utils import db
 
 imap_hook_string = 'airflow.contrib.sensors.imap_attachment_sensor.ImapHook'
@@ -33,7 +34,7 @@ class TestImapAttachmentSensor(unittest.TestCase):
     def setUp(self):
         configuration.load_test_config()
         db.merge_conn(
-            models.Connection(
+            Connection(
                 conn_id='imap_test',
                 host='base_url',
                 login='user',
