+++ tests/contrib/sensors/test_qubole_sensor.py
@@ -25,7 +25,8 @@ from mock import patch
 
 from airflow.contrib.sensors.qubole_sensor import QuboleFileSensor, QubolePartitionSensor
 from airflow.exceptions import AirflowException
-from airflow.models import DAG, Connection
+from airflow.models import DAG
+from airflow.models.connection import Connection
 from airflow.utils import db
 
 DAG_ID = "qubole_test_dag"
