+++ tests/hooks/test_docker_hook.py
@@ -20,8 +20,8 @@
 import unittest
 
 from airflow import configuration
-from airflow import models
 from airflow.exceptions import AirflowException
+from airflow.models.connection import Connection
 from airflow.utils import db
 
 try:
@@ -43,7 +43,7 @@ class DockerHookTest(unittest.TestCase):
     def setUp(self):
         configuration.load_test_config()
         db.merge_conn(
-            models.Connection(
+            Connection(
                 conn_id='docker_default',
                 conn_type='docker',
                 host='some.docker.registry.com',
@@ -52,7 +52,7 @@ class DockerHookTest(unittest.TestCase):
             )
         )
         db.merge_conn(
-            models.Connection(
+            Connection(
                 conn_id='docker_with_extras',
                 conn_type='docker',
                 host='some.docker.registry.com',
@@ -148,7 +148,7 @@ class DockerHookTest(unittest.TestCase):
 
     def test_conn_with_broken_config_missing_username_fails(self, _):
         db.merge_conn(
-            models.Connection(
+            Connection(
                 conn_id='docker_without_username',
                 conn_type='docker',
                 host='some.docker.registry.com',
@@ -165,7 +165,7 @@ class DockerHookTest(unittest.TestCase):
 
     def test_conn_with_broken_config_missing_host_fails(self, _):
         db.merge_conn(
-            models.Connection(
+            Connection(
                 conn_id='docker_without_host',
                 conn_type='docker',
                 login='some_user',
