+++ tests/hooks/test_mysql_hook.py
@@ -24,8 +24,8 @@ import unittest
 
 import MySQLdb.cursors
 
-from airflow import models
 from airflow.hooks.mysql_hook import MySqlHook
+from airflow.models.connection import Connection
 
 SSL_DICT = {
     'cert': '/tmp/client-cert.pem',
@@ -39,7 +39,7 @@ class TestMySqlHookConn(unittest.TestCase):
     def setUp(self):
         super(TestMySqlHookConn, self).setUp()
 
-        self.connection = models.Connection(
+        self.connection = Connection(
             login='login',
             password='password',
             host='host',
